chore: moved App.kt and renamed it to ApplicationCallExt.kt to keep history
