+++ ebms-provider/src/main/kotlin/no/nav/emottak/ebms/ebxml/ApplicationCallExt.kt
@@ -1,7 +1,7 @@
 /*
  * This Kotlin source file was generated by the Gradle 'init' task.
  */
-package no.nav.emottak.ebms
+package no.nav.emottak.ebms.ebxml
 
 import dev.reformator.stacktracedecoroutinator.runtime.DecoroutinatorRuntime
 import io.ktor.client.request.forms.MultiPartFormDataContent
@@ -38,9 +38,20 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import net.logstash.logback.marker.Markers
 import no.nav.emottak.constants.SMTPHeaders
+import no.nav.emottak.ebms.CpaRepoClient
+import no.nav.emottak.ebms.EBMS_PAYLOAD_SCOPE
+import no.nav.emottak.ebms.EBMS_SEND_IN_SCOPE
+import no.nav.emottak.ebms.PayloadProcessingClient
+import no.nav.emottak.ebms.SendInClient
+import no.nav.emottak.ebms.defaultHttpClient
 import no.nav.emottak.ebms.model.DokumentType
 import no.nav.emottak.ebms.model.EbMSDocument
+import no.nav.emottak.ebms.navCheckStatus
+import no.nav.emottak.ebms.postEbmsAsync
+import no.nav.emottak.ebms.postEbmsSync
 import no.nav.emottak.ebms.processing.ProcessingService
+import no.nav.emottak.ebms.registerHealthEndpoints
+import no.nav.emottak.ebms.scopedAuthHttpClient
 import no.nav.emottak.ebms.sendin.SendInService
 import no.nav.emottak.ebms.validation.DokumentValidator
 import no.nav.emottak.ebms.validation.MimeHeaders
@@ -61,6 +72,17 @@ import java.time.Instant
 import java.util.Base64
 import java.util.UUID
 import kotlin.time.toKotlinDuration
+import no.nav.emottak.ebms.CpaRepoClient
+import no.nav.emottak.ebms.EBMS_PAYLOAD_SCOPE
+import no.nav.emottak.ebms.EBMS_SEND_IN_SCOPE
+import no.nav.emottak.ebms.PayloadProcessingClient
+import no.nav.emottak.ebms.SendInClient
+import no.nav.emottak.ebms.defaultHttpClient
+import no.nav.emottak.ebms.navCheckStatus
+import no.nav.emottak.ebms.postEbmsAsync
+import no.nav.emottak.ebms.postEbmsSync
+import no.nav.emottak.ebms.registerHealthEndpoints
+import no.nav.emottak.ebms.scopedAuthHttpClient
 
 val log = LoggerFactory.getLogger("no.nav.emottak.ebms.App")
